Drop include of immintrin.h from F14Table.h
That include file ended up including a definition for a _serialize()
macro starting in GCC 11, which conflicted with Python definitions in
fbthrift and broke that build.

That file was originally included for the __m128i definition, but that
type definition actually comes from emmintrin.h, which is already
included.

Tested that this RPM package still builds correctly, and building a new
fbthrift on top of this change works successfully without any patches to
rename the _serialize() symbol.

Signed-off-by: Filipe Brandenburger <filbranden@gmail.com>